update: moved tests to test.rs and cleaned the server and header.rs.
